+++ packages/alias/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ## Breaking Changes
 
-- feat(alias): initial implementation of alias [bd6c230e](https://github.com/tylors/ragnar/commits/bd6c230e1f74cf83946df191206b10debe5ce92b)
+- feat(alias): initial implementation of alias [bd6c230e](https://github.com/tylors/reginn/commits/bd6c230e1f74cf83946df191206b10debe5ce92b)