+++ packages/app/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ## Breaking Changes
 
-- feat(app): Applications for Ragnar [43664b4f](https://github.com/tylors/ragnar/commits/43664b4f025678ae5b9c31cf5ce88d793549cba8)
+- feat(app): Applications for Reginn [43664b4f](https://github.com/tylors/reginn/commits/43664b4f025678ae5b9c31cf5ce88d793549cba8)