+++ packages/callback/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ## Breaking Changes
 
-- feat(callback): Add a callback to matched commands [3b94b099](https://github.com/tylors/ragnar/commits/3b94b0990df2740c8297ac3e3f28e84360344b29)
+- feat(callback): Add a callback to matched commands [3b94b099](https://github.com/tylors/reginn/commits/3b94b0990df2740c8297ac3e3f28e84360344b29)