+++ packages/command/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ## Breaking Changes
 
-- feat(command): Commands for Ragnar [96cde3c6](https://github.com/tylors/ragnar/commits/96cde3c61991bd20676277ebb624c61aa3fbe0fd)
+- feat(command): Commands for Reginn [96cde3c6](https://github.com/tylors/reginn/commits/96cde3c61991bd20676277ebb624c61aa3fbe0fd)