+++ packages/desc/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ## Breaking Changes
 
-- feat(desc): Add descriptions to types [d96eccbf](https://github.com/tylors/ragnar/commits/d96eccbfe614dfd28b45899f4dff056eebb417a3)
+- feat(desc): Add descriptions to types [d96eccbf](https://github.com/tylors/reginn/commits/d96eccbfe614dfd28b45899f4dff056eebb417a3)