+++ packages/help/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ## Breaking Changes
 
-- feat(help): Add help command to your application [927be178](https://github.com/tylors/ragnar/commits/927be17885609fe5c68e6e70eca1c761ccd64b43)
+- feat(help): Add help command to your application [927be178](https://github.com/tylors/reginn/commits/927be17885609fe5c68e6e70eca1c761ccd64b43)