+++ packages/promise/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ## Breaking Changes
 
-- feat(promise): Handle commands with a promise [8e9aa458](https://github.com/tylors/ragnar/commits/8e9aa458292d8b4f17b039857da29030df155f16)
+- feat(promise): Handle commands with a promise [8e9aa458](https://github.com/tylors/reginn/commits/8e9aa458292d8b4f17b039857da29030df155f16)