+++ packages/stream/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ## Breaking Changes
 
-- feat(stream): Handle your commands with a stream [2b65817b](https://github.com/tylors/ragnar/commits/2b65817bb924f6a6880da44398f0f32f0f00139b)
+- feat(stream): Handle your commands with a stream [2b65817b](https://github.com/tylors/reginn/commits/2b65817bb924f6a6880da44398f0f32f0f00139b)