+++ packages/type/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ## Breaking Changes
 
-- feat(type): change how flags are parsed [90875b18](https://github.com/tylors/ragnar/commits/90875b18a28a32da48ff44d99deaffab9bfd926d)
+- feat(type): change how flags are parsed [90875b18](https://github.com/tylors/reginn/commits/90875b18a28a32da48ff44d99deaffab9bfd926d)